refactor Regex a bit
