+++ iina/AppData.swift
@@ -26,10 +26,6 @@ struct AppData {
   /** generate aspect and crop options in menu */
   static let aspects: [String] = ["4:3", "5:4", "16:9", "16:10", "1:1", "3:2", "2.21:1", "2.35:1", "2.39:1"]
 
-  static let aspectRegex = Utility.Regex("\\A\\d+(\\.\\d+)?:\\d+(\\.\\d+)?\\Z")
-  static let HTTPFileNameRegex = Utility.Regex("attachment; filename=(.+?)\\Z")
-  static let tagVersionRegex = Utility.Regex("\\Av(\\d+\\.\\d+\\.\\d+)\\Z")
-
   static let rotations: [Int] = [0, 90, 180, 270]
 
   /** Seek amount */
